Parser now matches block intervals to price intervals
This is the initial implementation of the newer parsing model. The overriding goal here is to make the twitter blocks match our price intervals more closely. The current implementation will use the price data's intervals for the tweet blocks. The next iteration of the class will match up tweet blocks with price data by date