Fixed a bug with initialPrice assingement
In the for loop beginning on line 7, initialprice was referenced as the first element of a tuple. This worked well on the first iteration, where the initialprice variable was assigned as the first tuple in the USD list. However, on the second iteration initialprice was reassigned as a float, and when getChange tried to access the second index of initialprice it was actually trying to access  the second integer in the float. This is a hidden danger of python's mutable types!